backend/TransactionAnalysisService/requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
